+++ src/plugins/blocksmith/components/instagram-media-block-modal/index.jsx
@@ -151,7 +151,7 @@ export default class InstagramMediaBlockModal extends React.Component {
               onChange={this.handleChangeTextarea}
               placeholder="enter url or embed code"
               type="textarea"
-              value={displayUrl || null}
+              value={displayUrl || ''}
             />
           </FormGroup>
           {
@@ -185,10 +185,7 @@ export default class InstagramMediaBlockModal extends React.Component {
                 />
               </div>
             )}
-          {
-            isValid
-            && <InstagramMediaBlockPreview block={this.setBlock()} />
-          }
+          {isValid && <InstagramMediaBlockPreview block={this.setBlock()} />}
         </ModalBody>
         <ModalFooter>
           <Button onClick={toggle}>Cancel</Button>