+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## vN.N.N
 * Do not update active edits screen unless user clicks "refresh" button.
+* Instagram embed preview not rendering.
 * Gallery Launch Text not reflecting with red background.
 * Set user node title using first and last name fields in the UserSubscriber to avoid an inconsistency.